+++ tests/test_utilities.py
@@ -159,19 +159,19 @@ def test_empty_lonlat_file():
 
 def test_error_if_no_filename():
     with pytest.raises(Exception) as exc_info:
-        utilities.points_from_spatial("")
+        utilities.points_from_spatial("", "fake_gsr")
     assert re.search("spatial_dir is specified but no", exc_info.value.args[0])
 
 
 def test_reverses_closed_spo_points():
     lonlats = utilities.raw_points("./fixtures/spatial/closed.spo")
-    spo_lonlats = utilities.parse_spo("./fixtures/spatial/closed.spo")
+    spo_lonlats = utilities.parse_spo(constants.GEODETIC, lonlats)
     assert lonlats[1] == spo_lonlats[-2]
 
 
 def test_reverses_open_spo_points():
     lonlats = utilities.raw_points("./fixtures/spatial/open.spo")
-    spo_lonlats = utilities.parse_spo("./fixtures/spatial/open.spo")
+    spo_lonlats = utilities.parse_spo(constants.GEODETIC, lonlats)
     assert lonlats[1] == spo_lonlats[-2]
     assert len(lonlats) == len(spo_lonlats) - 1
 
@@ -188,3 +188,10 @@ def test_accepts_closed_polygon(closed_polygon):
 
 def test_ignores_tiny_spo(not_a_polygon):
     assert utilities.closed_polygon(not_a_polygon) == not_a_polygon
+
+
+def test_spo_is_geodetic():
+    with pytest.raises(Exception):
+        utilities.parse_spo(
+            constants.CARTESIAN, utilities.raw_points("./fixtures/spatial/open.spo")
+        )